Removed configurable dbdriver and dbport. update_local script can be used as installer for project

+++ scripts/_helpers/symfony.sh
@@ -4,7 +4,15 @@
 
 ## Symfony bootstrap
 # Ensure console symlink in bin
-ln -snf $ROOT/app/console $ROOT/bin/console
+
+function link_console() {
+    if [ -e $ROOT/bin ] ;
+    then
+        ln -snf $ROOT/app/console $ROOT/bin/console
+    fi
+}
+
+
 
 function delete_cache() {
     rm -rf $ROOT/app/cache/* $ROOT/app/logs/*
@@ -13,3 +21,6 @@ function delete_cache() {
 function composer_install_prod() {
     composer install --no-dev --optimize-autoloader --no-interaction
 }
+
+### run hooks
+link_console
